Changes:
  - Add environments path to kitchen yaml so it can find the environment
    files
  - Set the environment for kitchen test images to kitchen
  - Comment out the segulja_packages cookbook
  - Comment out the segulja_autofs cookbook
  - Add a check for the environment to the default recipe so that it
    will not automatically reboot kitchen test images

On branch new-server
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   .kitchen.yml
	modified:   metadata.rb
	modified:   recipes/default.rb

+++ recipes/default.rb
@@ -6,21 +6,23 @@
 
 include_recipe 'chef-client::default'
 include_recipe 'segulja_users::default'
-include_recipe 'segulja_packages::default'
+# include_recipe 'segulja_packages::default'
 include_recipe 'segulja_sudo::default'
-include_recipe 'segulja_autofs::default'
+# include_recipe 'segulja_autofs::default'
 include_recipe 'segulja_hostfile::default'
 
-execute 'Reboot after initial build' do
-  command 'shutdown -r 1'
-  action :run
-  not_if { File.exist?('/root/.initial-build-complete') }
-end
+if node.chef_environment != 'kitchen'
+  execute 'Reboot after initial build' do
+    command 'shutdown -r 1'
+    action :run
+    not_if { File.exist?('/root/.initial-build-complete') }
+  end
 
-file '/root/.initial-build-complete' do
-  content 'DO NOT DELETE'
-  owner 'root'
-  group 'root'
-  mode '0400'
-  action :create
+  file '/root/.initial-build-complete' do
+    content 'DO NOT DELETE'
+    owner 'root'
+    group 'root'
+    mode '0400'
+    action :create
+  end
 end
